snap: add support for building a classic snap package
Define a basic snapcraft.yaml to build Tilix using latest ldc and dub and
use a wrapper script where we set the minimum env variables in order to run.
Preloading of glib is needed as the rpath set by snapcraft will otherwise
load the core one instead that the one provided by the runtime.